Move getCurrentTimeMs into timer utils
