+++ src/common/timer.h
@@ -15,4 +15,6 @@ time_point get_server_start_time(void);
 void timer_init(void);
 void timer_final(void);
 
+uint32 getCurrentTimeMs();
+
 #endif
